MOD: 타입 export 파일 수정
